Setup "Is reviewing request" state
